[FEATURE] UNTESTED: Added FormConfig metadata through annotations
Added metadata classes for entities that are eligible for the use in creating FormConfiguration instances. The annotations are needed to be able to create a list of eligible entities and their fields (including data types) that can be sent to the client when a user uses the FormBuilder. This should improve type safety in FormConfiguration creation for the target entity's type, along with its fields.